+++ frame/common/src/ciphertexts/mod.rs
@@ -1,2 +1,1 @@
-pub mod enclave_key;
 pub mod treekem;
